backend/requirements.txt
fastapi
uvicorn[standard]
langchain==1.2.11
langchain-core==1.2.18
langgraph==1.1.0
langchain-google-genai==4.2.1
google-generativeai==0.8.6
google-genai==1.66.0
sqlalchemy
pydantic
pydantic-settings
python-multipart
httpx
python-jose[cryptography]
passlib[bcrypt]
watchfiles
psycopg2-binary
asyncpg
greenlet
pgvector
firebase-admin
google-cloud-storage
python-dotenv
openai
aiohttp
aiofiles
pypdf
python-docx
python-pptx
google-api-python-client
cryptography
langchain-groq
beautifulsoup4
lxml
groq
langchain-google-vertexai
google-cloud-aiplatform
google-cloud-tasks
google-auth
google-auth-oauthlib
google-auth-httplib2
slowapi==0.1.9
stripe
openpyxl
mammoth

matplotlib==3.8.4
Pillow
moviepy==1.0.3
